v3: Define & Generate API responses from Perl
This allows Perl modules to define new API responses and the
corresponding Elm type and parser will be generated automatically. The
primary reason to do this is to ensure that Elm and Perl stay
synchronized and to make sure that errors in generating API responses
can easily be found on the server side.

Also moded json_api() from VN3::Validate to VN3::Prelude, as that seemed
more fitting.