v2rw: Add Elm & db_edit framework + Convert doc page editing
Most of this is copied from v3. I did improve on a few aspects:

- db_edit() and db_entry() use VNDB::Schema rather than dynamically
  querying the DB. This has the minor advantage of a faster startup.
- The Elm code generator now writes to multiple files, this avoids
  the namespace pollution seen in v3's Lib.Gen and makes the dependency
  graph a bit more lean (i.e. faster incremental builds).
- The Elm code generator doesn't update the timestamp of files that
  haven't been modified. This also speeds up incremental builds, the elm
  compiler can now skip rebuilding unmodified files.
- The Elm API response generator code now uses plain functions rather
  than code references and all possible responses are now defined in
  Elm.pm. Turns out most API responses were used from more than a single
  place, so it makes sense to have them centrally defined.

The doc page preview function is also much nicer; I'd like to apply this
to all BBCode textareas as well.

(Elm.pm itself is ugly as hell though. And we will prolly need some HTML
form generation functions in Elm to make that part less verbose)