JS: Add LibreJS compatibility + use embedded comment tag to sort v2rw.js files
I found myself forgetting to add a file to $JS_FILES in the Makefile
several times now, and with Perl and Elm modules being detected and
loaded automatically, it feels weird not to do the same for .js files.
This embedded order trick is ugly, but it's pretty simple and gets the
job done.