JS: Remove old "This is not VNDB!" domain check
This is a remnant from when I was running a public sandbox VNDB, to
avoid getting people confused and wasting their time contributing to the
sandbox.

Removed for consistency with pages that have migrated to the new
v2rw.js, which doesn't have the message, and for the new .onion address.