ulist: Get rid of BETA warning, update menu and tab links
Decided to go with option presets (e.g. '?wishlist=1') rather than add
the full list of options to each link. Makes it easier to maintain and
keeps URLs for basic pages short.

As a result of that, I had to fix ColSelect.elm to not read the current
URL but instead get the URL passed by Perl. Fortunately, it looks like
the Erl package can deal with partial URLs (only the query string, in
this case) just fine.