ulist: Rework VN "User options" tab to be more prominent and convenient
Personally not a big fan of this taking so much space when it's not on
the user's list, but people want instant options so there we go. It
probably still blends in too well with the rest of the VN info though,
styling isn't my thing.

And there's no need to hit "Add to list" anymore now.