Docker: Switch to Alpine
Main reason is because I'm more familiar with it, it has less weird
surprises than Ubuntu (well, apart from musl + busybox, but that only
helps increase VNDB's portability) and it's easier to update to later
software versions. Current image comes with PostgreSQL 12 and Perl 5.30,
which I will no doubt want to use in the future.