v2rw: WIP: Convert character pages
This uses an alternative (non-JS) approach to spoiler hiding, which is
both much simpler to implement and properly handles all cases. Downside
is that it requires a page reload. It also allows direct linking to a
character page with a particular view, but I'm not entirely sure if this
is a welcome feature.

This is a WIP because it doesn't display instances yet.