v2rw: Add instance listing to Chars::Page
I wasn't sure how to handle the fetching of multiple character entries;
whether to extend/re-use db_entry() or write a separate function. Ended
up going with the latter solution, as db_entry() already has other
functionality (and overhead) that we won't be needing in this case. This
overhead will matter when this function is used for the character
listing on VN pages.