Chars::Page: Only show spoiler & sexual traits settings when they do something
i.e. when there's something on the page that would be hidden with
different settings.

For individual fields, just knowing that something is a spoiler is in
itself a spoiler. But when it comes to the entire page, I can't imagine
a case where "there is a spoiler somewhere on this page" would by itself
be a spoiler, so this should be safe enough.

Better than clicking on spoiler settings that don't do anything.