v2rw/WIP: Incremental rewrite of the release edit form
The new form does change a few things to the UI, so I'll keep the old
form around for a while so that people have something to fallback to
when it doesn't work (like, right now).

I'm also doing this rewrite in multiple commits as it's a pretty large
change.