v2rw/RelEdit: Add external links editor
Seems to be working pretty well. It's not as discoverable as the old
form, but it's much easier to work with (just paste the URL), it
provides better feedback and looks a lot less messy.

The code is also nicely DRY. It has some complexity, but less than what
I had anticipated. Adding new links or allowing multiple links for a
single site should be much easier now.

This is the edit form I was referring to back in
06bacb61526f3945520dd344821d2aa7b85a5f43