v2rw/RelEdit: Allow creating a new release
Not totally super happy with this solution; I'd rather automatically
create an empty entry and send that to 'RelEdit.Main', but initializing
all fields in Perl is tricky. At least in Elm we get some sort of type
checking (though Elm can't tell whether the default value makes sense)
and doing this in Elm makes it possible to initialize to something
invalid that the user has to change (not currently done yet).

It's messy either way.