refactor: Add elm_empty() function to initialize a new form
This gets rid of the separate Elm-based empty form initialization and
the ReleaseEdit.New module. There's rather less repitition with this
approach.

This is an alternative to the approach I implemented in
255dc704f4675326ebf6358c79b5b37fca903e4b - there I mentioned creating
such a data structure in Perl as "tricky", which I thought it was,
except it turns out to be pretty easy to automate in a reliable way.

I'll convert the StaffEdit form to use this approach as well, after
testing this approach a bit more.