imgflag: Disallow regular users to vote on any image of their choosing
And reduce the number of images given out randomly. That reduction makes
it more expensive for an abuser to get tokens for the images they want
(they'd hit the req/sec rate limiter much earlier), and also makes sense
now that the query has been optimized with the TABLESAMPLE trick (it's
faster, and more queries with less rows ought to give a better
distribution overall - it reduces the impact of a biased query).