imgflag: Use onKeyDown/Up in tandem to provide feedback over selected option
Hopefully less error-prone. Ctrl+L now also doesn't select "brutal"
anymore, thanks to the KeyUp requirement.

I experimented with moving to the next image with a timer (100-200ms),
but any form of delay is pretty annoying.

This would be even nicer if there was a way to cancel the KeyUp event
if you decide it was the wrong selection.