imgflag: Adjust weight algorithm to more heavily bias towards low vote counts
Using an exponential rather than linear function. I also found that a
single outlier vote has a pretty high influence on the stddev, so I've
attenuated that further by taking (stddev/max_stddev)^2. (which happens
to correspond to the variance because out stddev is always between 0 and
1)

Due to the exponential bias on vote count, it now takes 7 or 8 votes
before the influence of the stddev becomes very relevant. Which means
that, with the current vote counts, the stddev doesn't matter much at
all.