v2rw/VN::Page: Add staff listing
The old staff organization was such a chaotic mess of unused space that
I figured I could improve on that with some proper bin packing.
Unfortunately, doing that well in CSS proved impossible, so I went with
generating three layouts on the server and using @media queries to
show/hide the appropriate one based on browser width. This bloats the
HTML a bit, but oh well...

This new organization significantly reduces the vertical space used for
staff on most pages, so I took that as a chance to get rid of the
summarization (the "more" link). With smaller browser widths or very
long staff listings that may still be necessary though, so maybe it'll
come back.