v2rw/VN::Page: Add character summary + mainbox summarization
Didn't see a better way to keep the height of the character summary
in check, so re-added the mainbox summarization.

Spoiler hiding now follows user preferences - this fixes a correctness
issue with characters being assigned the wrong role if they have
spoilerous release-specific roles and is also just a lot easier.
Replaced the spoiler options with a link to the full character list
instead.