v2rw/VN::Page: Converted screenshots + image viewer
Decided to base the image viewer on the old one, rather than backport
the Lightbox from v3. That Lightbox wasn't very polished yet and had a
few remaining usability issues. Additionally, Elm doesn't have a whole
lot to offer for a simple image viewer: the complexity does not lie in
state management and view updates but rather in the interaction with the
rest of the page.

A few changes compared to the old one:
- Esc closes the viewer
- position: fixed
- Position and dimensions of the viewer are dynamically updated when the
  window resizes.
- Clicking on a link outside of the viewer now performs the action of
  that link (i.e. no more preventDefault()).
- Less ugly loading indicator.