Drop ReleaseExtLinks.elm in favor of pure CSS solution
It's the only Elm module used on VN pages for not-logged-in visitors, so
doing this in CSS makes it possible to not load an Elm bundle on those
pages (when I get to splitting up the JS files).

This solution also simplifies code reuse with the producer pages, but a
v2rw migration of that page would have taken care of that as well.