v2rw: Convert producer pages & consolidate release row HTML
The release listing on producer pages now uses the same HTML as on VN
pages, so it comes with the same information & features. Unfortunately,
it also comes with the same messiness...

The page is kind of large and slow for producers with lots of releases,
not sure what to do with that.