VN::Edit: Add length/wikidata/renai/anime fields
Completing the "General info" tab. This makes use of the new anime
titles import for validation and autocomplete.

It does have a bug, though, as it also autocompletes deleted AniDB
entries that happen to have been linked to old VN revisions. Need to
find a way to mark those for exclusion.