VN::Page: Display screenshot flagging status in imageviewer thing
Replacing the 'close' button with something that does pretty much the
opposite may not be the wisest thing to do, but that close button has
always been a bit useless considering that clicking almost anywhere on
the page will have the same effect. The right-bottom of the image is the
natural place for the flagging status, as that's also where the status
is displayed for other images on the site.