Reports: Fix display of deleted threads
This actually hides a bigger problem: There's no way to uniquely link to
forum posts when posts can be deleted, because the (tid,num) identifier
will be re-used for later posts and the report will then point to a
completely different post.

Not trivial to fix, and hopefully not too much of a problem when you're
aware of it.