Discussions: Fix handling of user deletion
Which I broke by removing the DEFAULT clause on threads_posts.uid. That
column still used the old uid=0 for deleted users rather than the
uid=NULL that I was planning to migrate the entire DB to. So while I'm
fixing up the threads schema, may as well update this too.

There's still a bunch of columns relying on uid=0, but I can fix that
later.