Reports: Get rid of report_type, split vndbid/num + allow reporting reviews and comments
rtype is not necessary, the DB identifier is sufficient. The separate
type column was supposed to simplify DB lookups (but that's all
consolidated in a single function anyway, so it didn't help much) and
easy filtering depending on what mods have access to (but there is no
such filter, and even with vndbids that should be easy enough).

Converting the object ID into a vndbid + num should make it easier to
correlate reports if necessary.